Extract and document the animation
